+++ src/guile/ares/suitbl/definitions.scm
@@ -3,7 +3,7 @@
 
 (define-module (ares suitbl definitions)
   #:export (current-test-runner
-            set-current-test-runner!
+            set-default-test-runner!
 
             is
             test test?
@@ -29,8 +29,8 @@ library, which sets an approriate test runner for you."))
 
 (define current-test-runner (make-parameter missing-test-runner))
 
-(define (set-current-test-runner! runner)
-  "Set the current test runner to RUNNER and return the previous runner."
+(define (set-default-test-runner! runner)
+  "Set the default test runner to RUNNER and return the previous runner."
   (current-test-runner runner))
 
 (define (test? x)